fix: remove unneded type exports in mongodb instrumentation
Avoid leaking types from mongodb instrumentation by splitting types.ts into a public
and internal file.
Added type Document to avoid that lib dom needs to be added in build.
